Remove unnessary files
1. as title
2. create obsolete dir for temporary

+++ setup.sh
@@ -5,34 +5,35 @@
 set -x
 
 PWD=`pwd`
+ENV='env'
+ETC='etc'
+TOOLS='tools'
+
 SRC=(
-"$PWD/environment/bin"
-"$PWD/environment/vim/vimrc"
-"$PWD/environment/android"
-"/media/karl/ext4/Projects"
-"/media/karl/ntfs/MEGA"
+"$PWD/$ENV/bin"
+"$PWD/$ENV/vim/vimrc"
+"$PWD/$ENV/android"
 )
 
 GOAL=(
 "$HOME/bin"
 "$HOME/.vimrc"
 "$HOME/.android"
-"$HOME/Projects"
-"$HOME/MEGA"
 )
 
-echo ". $PWD/tools/env.cfg" >> $HOME/.profile
+SRC_AND_DEST=(
+"$PWD/$ENV/bin:$HOME/bin"
+"$PWD/$ENV/vim/vimrc:$HOME/.vimrc"
+"$PWD/$ENV/android:$HOME/.android"
+)
+
+echo ". $PWD/$TOOLS/env.cfg" >> $HOME/.profile
 
 i=0
-while [ true ]; do
-	if [ -z ${SRC[$i]} ]; then
-		break;
-	fi
-	if [ -e ${GOAL[$i]} ]; then
-		mv ${GOAL[$i]} ${GOAL[$i]}.bak
-	fi
+for i in $(0:${#SRC_AND_DEST[@]}) do
+        [ -z ${SRC[$i]} ] && continue;
+	[ -e ${GOAL[$i]} ] && mv ${GOAL[$i]} ${GOAL[$i]}.bak
 	cp ${SRC[$i]} ${GOAL[$i]}
-	i=$((i+1))
 done
 
 set +x
